Dry_Wet_Waste_Segregation
For implementing this hardware setup you will need to learn the following concepts in depth:

Arduino : Arduino is a microcontroller which will be used in this project. You can learn everything about this at this link . Also don't forget to check the official documentation of Arduino.
C++ Programming : For programming Arduino you will to learn C++ Programming to an intermediate level. You will also have to learn the Basic functions of Arduino.
Moisture Sensor : You will need to learn the Interfacing of Arduino with the Moisture sensor with circuit diagram.
Touch Sensor : You will need to learn the Interfacing of Arduino with the Touch sensor with circuit diagram.
Servo Motor : You will have to learn about Servo Motor and how exactly it is operated with Arduino.